requirements.txt
streamlit==1.45.1
langchain==0.1.12
langchain-community==0.0.27
langchain-google-genai==0.0.11
python-dotenv==1.0.1
PyMuPDF==1.23.26
faiss-cpu==1.8.0
google-generativeai==0.8.5
